Fix bug that causes views to update on initial window move
When the application is launched, and the window is moved by its title bar, all views gets updated. This commit initializes all newly added member variables, and sets 'm_IsResizing' to false on the first update, as this gets set to true during the initial 'resize' which is caused on window apparition post construction.